move managers to filer.managers, make File and Folder managers swappable via settings
