+++ filer/models/foldermodels.py
@@ -2,89 +2,24 @@
 
 from __future__ import absolute_import, unicode_literals
 
-import mptt
 from django.conf import settings
 from django.contrib.auth import models as auth_models
 from django.core import urlresolvers
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.db.models import Q
 from django.utils.http import urlquote
 from django.utils.translation import ugettext_lazy as _
+from mptt.models import MPTTModel
 
 from . import mixins
 from .. import settings as filer_settings
 from ..utils.compatibility import python_2_unicode_compatible
-
-
-class FolderManager(models.Manager):
-    def with_bad_metadata(self):
-        return self.get_query_set().filter(has_all_mandatory_data=False)
-
-
-class FolderPermissionManager(models.Manager):
-    """
-    Theses methods are called by introspection from "has_generic_permisison" on
-    the folder model.
-    """
-    def get_read_id_list(self, user):
-        """
-        Give a list of a Folders where the user has read rights or the string
-        "All" if the user has all rights.
-        """
-        return self.__get_id_list(user, "can_read")
-
-    def get_edit_id_list(self, user):
-        return self.__get_id_list(user, "can_edit")
-
-    def get_add_children_id_list(self, user):
-        return self.__get_id_list(user, "can_add_children")
-
-    def __get_id_list(self, user, attr):
-        if user.is_superuser or not filer_settings.FILER_ENABLE_PERMISSIONS:
-            return 'All'
-        allow_list = set()
-        deny_list = set()
-        group_ids = user.groups.all().values_list('id', flat=True)
-        q = Q(user=user) | Q(group__in=group_ids) | Q(everybody=True)
-        perms = self.filter(q).order_by('folder__tree_id', 'folder__level',
-                                        'folder__lft')
-        for perm in perms:
-            p = getattr(perm, attr)
-
-            if p is None:
-                # Not allow nor deny, we continue with the next permission
-                continue
-
-            if not perm.folder:
-                assert perm.type == FolderPermission.ALL
-
-                if p == FolderPermission.ALLOW:
-                    allow_list.update(Folder.objects.all().values_list('id', flat=True))
-                else:
-                    deny_list.update(Folder.objects.all().values_list('id', flat=True))
-
-                continue
-
-            folder_id = perm.folder.id
-
-            if p == FolderPermission.ALLOW:
-                allow_list.add(folder_id)
-            else:
-                deny_list.add(folder_id)
-
-            if perm.type == FolderPermission.CHILDREN:
-                if p == FolderPermission.ALLOW:
-                    allow_list.update(perm.folder.get_descendants().values_list('id', flat=True))
-                else:
-                    deny_list.update(perm.folder.get_descendants().values_list('id', flat=True))
-
-        # Deny has precedence over allow
-        return allow_list - deny_list
+from ..utils.loader import load_object
+from .managers import FolderPermissionManager
 
 
 @python_2_unicode_compatible
-class Folder(models.Model, mixins.IconsMixin):
+class Folder(MPTTModel, mixins.IconsMixin):
     """
     Represents a Folder that things (files) can be put into. Folders are *NOT*
     mirrored in the Filesystem and can have any unicode chars as their name.
@@ -112,7 +47,7 @@ class Folder(models.Model, mixins.IconsMixin):
     created_at = models.DateTimeField(_('created at'), auto_now_add=True)
     modified_at = models.DateTimeField(_('modified at'), auto_now=True)
 
-    objects = FolderManager()
+    objects = load_object(filer_settings.FILER_FOLDER_MANAGER)()
 
     @property
     def file_count(self):
@@ -234,12 +169,6 @@ class Folder(models.Model, mixins.IconsMixin):
         app_label = 'filer'
         verbose_name = _("Folder")
         verbose_name_plural = _("Folders")
-
-# MPTT registration
-try:
-    mptt.register(Folder)
-except mptt.AlreadyRegistered:
-    pass
 
 
 @python_2_unicode_compatible
